Preformance boost still needs some work

+++ src/main/java/wurmatron/voidrpg/api/cube/IReactor.java
@@ -5,6 +5,6 @@ public interface IReactor {
 		/**
 		 * Amount of energy this reactor can control at once
 		 */
-		double getMaxPower();
+		int getMaxPower();
 
 }
