+++ src/main/java/wurmatron/voidrpg/common/cube/special/energy/CubeReactorI.java
@@ -75,7 +75,7 @@ public class CubeReactorI implements ICube, IReactor {
 		}
 
 		@Override
-		public double getMaxPower () {
+		public int getMaxPower () {
 				return 5;
 		}
 }
