+++ src/main/java/wurmatron/voidrpg/common/cube/special/energy/CubeReactorII.java
@@ -75,7 +75,7 @@ public class CubeReactorII implements ICube,IReactor {
 		}
 
 		@Override
-		public double getMaxPower () {
+		public int getMaxPower () {
 				return 20;
 		}
 }
