+++ src/main/java/wurmatron/voidrpg/common/cube/special/energy/CubeReactorIII.java
@@ -75,7 +75,7 @@ public class CubeReactorIII implements ICube,IReactor {
 		}
 
 		@Override
-		public double getMaxPower () {
+		public int getMaxPower () {
 				return 100;
 		}
 }
